Fixed WordleCpp and changed dir structure
- WordleCpp was not properly building the user data string
- Changed WordleCpp DatabaseEntry to use template specialization
- src/(cpp, include) -> src, include; same with CWordle

+++ WordleCpp/src/console_app/console_app.cpp
@@ -44,7 +44,7 @@ static int SaveUsernames(std::set<std::string> &usernames, std::string const &fi
 
 void RunConsoleApp(std::unordered_set<std::string> const &dictionary, std::set<std::string> &usernames)
 {
-    srand(time(NULL));
+    srand((uint32_t)time(NULL));
 
     ProgramState state = ProgramState::LogIn;
     uint8_t runProgram = 1;
